monitor_app/requirements.txt
pyserial>=3.5
PyQt5>=5.15.10
pyqtgraph>=0.13.3
